+++ app/pages/AboutPage2/index.ts
@@ -1,0 +1,1 @@
+export * from "./AboutPage";
